# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "Event": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "date_earliest": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Earliest"
          },
          "date_latest": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Latest"
          },
          "date_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Value"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/Participant"
            },
            "title": "Participants",
            "type": "array"
          },
          "place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place"
          },
          "place_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Id"
          },
          "sensitivity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Sensitivity"
              },
              {
                "type": "null"
              }
            ]
          },
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "space_id",
          "type",
          "date_value",
          "date_earliest",
          "date_latest",
          "place_id",
          "place",
          "description",
          "sensitivity",
          "participants",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "Event",
        "type": "object"
      },
...
      "EventCreate": {
        "additionalProperties": false,
        "properties": {
          "date_value": {
            "anyOf": [
              {
                "maxLength": 100,
                "pattern": "^[A-Z0-9_]+( [A-Z0-9_]+)*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Value"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/ParticipantIn"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Participants",
            "type": "array"
          },
          "place_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Id"
          },
          "sensitivity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Sensitivity"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to `religion` for sacraments and `health` for medical events."
          },
          "type": {
            "pattern": "^[a-z][a-z0-9_]{1,49}$",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "participants"
        ],
        "title": "EventCreate",
        "type": "object"
      },
      "EventPatch": {
        "additionalProperties": false,
        "properties": {
          "date_value": {
            "anyOf": [
              {
                "maxLength": 100,
                "pattern": "^[A-Z0-9_]+( [A-Z0-9_]+)*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Value"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "participants": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ParticipantIn"
                },
                "maxItems": 50,
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participants"
          },
          "place_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Id"
          },
          "sensitivity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Sensitivity"
              },
              {
                "type": "null"
              }
            ]
          },
          "type": {
            "anyOf": [
              {
                "pattern": "^[a-z][a-z0-9_]{1,49}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "title": "EventPatch",
        "type": "object"
      },
...
      "Participant": {
        "properties": {
          "person_id": {
            "format": "uuid",
            "title": "Person Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/ParticipantRole"
          }
        },
        "required": [
          "person_id",
          "role"
        ],
        "title": "Participant",
        "type": "object"
      },
      "ParticipantIn": {
        "additionalProperties": false,
        "properties": {
          "person_id": {
            "format": "uuid",
            "title": "Person Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/ParticipantRole"
          }
        },
        "required": [
          "person_id",
          "role"
        ],
        "title": "ParticipantIn",
        "type": "object"
      },
      "ParticipantRole": {
        "enum": [
          "principal",
          "spouse",
          "parent",
          "godparent",
          "witness",
          "officiant",
          "informant",
          "other"
        ],
        "title": "ParticipantRole",
        "type": "string"
      },
...
      "Sensitivity": {
        "enum": [
          "religion",
          "health",
          "genetic",
          "ethnicity",
          "sexual",
          "political"
        ],
        "title": "Sensitivity",
        "type": "string"
      },
...
    }
  },
  "info": {
    "summary": "Evidence-based family trees and family memory. AGPL-3.0-only.",
    "title": "family-history API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/events/{event_id}": {
      "delete": {
        "operationId": "delete_event_v1_events__event_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Delete Event",
        "tags": [
          "events"
        ]
      },
      "patch": {
        "operationId": "update_event_v1_events__event_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Update Event",
        "tags": [
          "events"
        ]
      }
    },
...
    "/v1/spaces/{space_id}/events": {
      "post": {
        "description": "Sacraments default to the `religion` sensitivity and medical events to `health`.",
        "operationId": "create_event_v1_spaces__space_id__events_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Create Event",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}